requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt
gunicorn
pillow
markdown
python-telegram-bot
gunicorn
mysqlclient
discord.py
python-dotenv